+++ governance-20210120/alibabacloud_governance20210120/models.py
@@ -526,6 +526,39 @@ class EnrollAccountRequestBaselineItems(TeaModel):
         return self
 
 
+class EnrollAccountRequestTag(TeaModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+    ):
+        self.key = key
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.key is not None:
+            result['Key'] = self.key
+        if self.value is not None:
+            result['Value'] = self.value
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+        return self
+
+
 class EnrollAccountRequest(TeaModel):
     def __init__(
         self,
@@ -538,6 +571,7 @@ class EnrollAccountRequest(TeaModel):
         payer_account_uid: int = None,
         region_id: str = None,
         resell_account_type: str = None,
+        tag: List[EnrollAccountRequestTag] = None,
     ):
         # The prefix for the account name of the member.
         # 
@@ -581,6 +615,190 @@ class EnrollAccountRequest(TeaModel):
         # 
         # > This parameter is available only for resellers at the international site (alibabacloud.com).
         self.resell_account_type = resell_account_type
+        self.tag = tag
+
+    def validate(self):
+        if self.baseline_items:
+            for k in self.baseline_items:
+                if k:
+                    k.validate()
+        if self.tag:
+            for k in self.tag:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.account_name_prefix is not None:
+            result['AccountNamePrefix'] = self.account_name_prefix
+        if self.account_uid is not None:
+            result['AccountUid'] = self.account_uid
+        if self.baseline_id is not None:
+            result['BaselineId'] = self.baseline_id
+        result['BaselineItems'] = []
+        if self.baseline_items is not None:
+            for k in self.baseline_items:
+                result['BaselineItems'].append(k.to_map() if k else None)
+        if self.display_name is not None:
+            result['DisplayName'] = self.display_name
+        if self.folder_id is not None:
+            result['FolderId'] = self.folder_id
+        if self.payer_account_uid is not None:
+            result['PayerAccountUid'] = self.payer_account_uid
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+        if self.resell_account_type is not None:
+            result['ResellAccountType'] = self.resell_account_type
+        result['Tag'] = []
+        if self.tag is not None:
+            for k in self.tag:
+                result['Tag'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AccountNamePrefix') is not None:
+            self.account_name_prefix = m.get('AccountNamePrefix')
+        if m.get('AccountUid') is not None:
+            self.account_uid = m.get('AccountUid')
+        if m.get('BaselineId') is not None:
+            self.baseline_id = m.get('BaselineId')
+        self.baseline_items = []
+        if m.get('BaselineItems') is not None:
+            for k in m.get('BaselineItems'):
+                temp_model = EnrollAccountRequestBaselineItems()
+                self.baseline_items.append(temp_model.from_map(k))
+        if m.get('DisplayName') is not None:
+            self.display_name = m.get('DisplayName')
+        if m.get('FolderId') is not None:
+            self.folder_id = m.get('FolderId')
+        if m.get('PayerAccountUid') is not None:
+            self.payer_account_uid = m.get('PayerAccountUid')
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+        if m.get('ResellAccountType') is not None:
+            self.resell_account_type = m.get('ResellAccountType')
+        self.tag = []
+        if m.get('Tag') is not None:
+            for k in m.get('Tag'):
+                temp_model = EnrollAccountRequestTag()
+                self.tag.append(temp_model.from_map(k))
+        return self
+
+
+class EnrollAccountShrinkRequestBaselineItems(TeaModel):
+    def __init__(
+        self,
+        config: str = None,
+        name: str = None,
+        skip: bool = None,
+        version: str = None,
+    ):
+        # The configurations of the baseline item.
+        self.config = config
+        # The name of the baseline item.
+        self.name = name
+        # Specifies whether to skip the baseline item. Valid values:
+        # 
+        # *   false: The baseline item is not skipped.
+        # *   true: The baseline item is skipped.
+        self.skip = skip
+        # The version of the baseline item.
+        self.version = version
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.config is not None:
+            result['Config'] = self.config
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.skip is not None:
+            result['Skip'] = self.skip
+        if self.version is not None:
+            result['Version'] = self.version
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Config') is not None:
+            self.config = m.get('Config')
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('Skip') is not None:
+            self.skip = m.get('Skip')
+        if m.get('Version') is not None:
+            self.version = m.get('Version')
+        return self
+
+
+class EnrollAccountShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        account_name_prefix: str = None,
+        account_uid: int = None,
+        baseline_id: str = None,
+        baseline_items: List[EnrollAccountShrinkRequestBaselineItems] = None,
+        display_name: str = None,
+        folder_id: str = None,
+        payer_account_uid: int = None,
+        region_id: str = None,
+        resell_account_type: str = None,
+        tag_shrink: str = None,
+    ):
+        # The prefix for the account name of the member.
+        # 
+        # *   If the account baseline is applied to an account that is newly created, you must configure this parameter.
+        # *   If the account baseline is applied to an existing account, you do not need to configure this parameter.
+        self.account_name_prefix = account_name_prefix
+        # The account ID.
+        # 
+        # *   If the account baseline is applied to an account that is newly created, you do not need to configure this parameter.
+        # *   If the account baseline is applied to an existing account, you must configure this parameter.
+        self.account_uid = account_uid
+        # The baseline ID.
+        # 
+        # If this parameter is left empty, the default baseline is used.
+        self.baseline_id = baseline_id
+        # An array that contains baseline items.
+        # 
+        # If this parameter is specified, the configurations of the baseline items are merged with the baseline of the specified account. The configurations of the same baseline items are subject to the configuration of this parameter. We recommend that you leave this parameter empty and configure the `BaselineId` parameter to specify an account baseline and apply the configuration of the account baseline to the account.
+        self.baseline_items = baseline_items
+        # The display name of the account.
+        # 
+        # *   If the account baseline is applied to an account that is newly created, you must configure this parameter.
+        # *   If the account baseline is applied to an existing account, you do not need to configure this parameter.
+        self.display_name = display_name
+        # The ID of the parent folder.
+        # 
+        # *   If the account baseline is applied to an account that is newly created, you need to specify a parent folder. If you do not configure this parameter, the account is created in the Root folder.
+        # *   If the account baseline is applied to an existing account, you do not need to configure this parameter.
+        self.folder_id = folder_id
+        # The ID of the billing account.
+        # 
+        # *   If the account baseline is applied to an account that is newly created, you need to specify a billing account. If you do not configure this parameter, the self-pay settlement method is used for the account.
+        # *   If the account baseline is applied to an existing account, you do not need to configure this parameter.
+        self.payer_account_uid = payer_account_uid
+        # The region ID.
+        self.region_id = region_id
+        # The identity type of the member. Valid values:
+        # 
+        # *   resell (default): The member is an account for a reseller. A relationship is automatically established between the member and the reseller. The management account of the resource directory must be used as the billing account of the member.
+        # *   non_resell: The member is not an account for a reseller. The member is an account that is not associated with a reseller. You can directly use the account to purchase Alibaba Cloud resources. The member is used as its own billing account.
+        # 
+        # > This parameter is available only for resellers at the international site (alibabacloud.com).
+        self.resell_account_type = resell_account_type
+        self.tag_shrink = tag_shrink
 
     def validate(self):
         if self.baseline_items:
@@ -614,6 +832,8 @@ class EnrollAccountRequest(TeaModel):
             result['RegionId'] = self.region_id
         if self.resell_account_type is not None:
             result['ResellAccountType'] = self.resell_account_type
+        if self.tag_shrink is not None:
+            result['Tag'] = self.tag_shrink
         return result
 
     def from_map(self, m: dict = None):
@@ -627,7 +847,7 @@ class EnrollAccountRequest(TeaModel):
         self.baseline_items = []
         if m.get('BaselineItems') is not None:
             for k in m.get('BaselineItems'):
-                temp_model = EnrollAccountRequestBaselineItems()
+                temp_model = EnrollAccountShrinkRequestBaselineItems()
                 self.baseline_items.append(temp_model.from_map(k))
         if m.get('DisplayName') is not None:
             self.display_name = m.get('DisplayName')
@@ -639,6 +859,8 @@ class EnrollAccountRequest(TeaModel):
             self.region_id = m.get('RegionId')
         if m.get('ResellAccountType') is not None:
             self.resell_account_type = m.get('ResellAccountType')
+        if m.get('Tag') is not None:
+            self.tag_shrink = m.get('Tag')
         return self
 
 
@@ -1116,6 +1338,39 @@ class GetEnrolledAccountResponseBodyInputsBaselineItems(TeaModel):
         return self
 
 
+class GetEnrolledAccountResponseBodyInputsTag(TeaModel):
+    def __init__(
+        self,
+        key: str = None,
+        value: str = None,
+    ):
+        self.key = key
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.key is not None:
+            result['Key'] = self.key
+        if self.value is not None:
+            result['Value'] = self.value
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+        return self
+
+
 class GetEnrolledAccountResponseBodyInputs(TeaModel):
     def __init__(
         self,
@@ -1125,6 +1380,7 @@ class GetEnrolledAccountResponseBodyInputs(TeaModel):
         display_name: str = None,
         folder_id: str = None,
         payer_account_uid: int = None,
+        tag: List[GetEnrolledAccountResponseBodyInputsTag] = None,
     ):
         # The prefix of the account name.
         self.account_name_prefix = account_name_prefix
@@ -1138,10 +1394,15 @@ class GetEnrolledAccountResponseBodyInputs(TeaModel):
         self.folder_id = folder_id
         # The ID of the settlement account.
         self.payer_account_uid = payer_account_uid
+        self.tag = tag
 
     def validate(self):
         if self.baseline_items:
             for k in self.baseline_items:
+                if k:
+                    k.validate()
+        if self.tag:
+            for k in self.tag:
                 if k:
                     k.validate()
 
@@ -1165,6 +1426,10 @@ class GetEnrolledAccountResponseBodyInputs(TeaModel):
             result['FolderId'] = self.folder_id
         if self.payer_account_uid is not None:
             result['PayerAccountUid'] = self.payer_account_uid
+        result['Tag'] = []
+        if self.tag is not None:
+            for k in self.tag:
+                result['Tag'].append(k.to_map() if k else None)
         return result
 
     def from_map(self, m: dict = None):
@@ -1184,6 +1449,11 @@ class GetEnrolledAccountResponseBodyInputs(TeaModel):
             self.folder_id = m.get('FolderId')
         if m.get('PayerAccountUid') is not None:
             self.payer_account_uid = m.get('PayerAccountUid')
+        self.tag = []
+        if m.get('Tag') is not None:
+            for k in m.get('Tag'):
+                temp_model = GetEnrolledAccountResponseBodyInputsTag()
+                self.tag.append(temp_model.from_map(k))
         return self
 
 
@@ -2676,6 +2946,7 @@ class ListEvaluationMetricDetailsResponseBodyResourcesResourceProperties(TeaMode
 class ListEvaluationMetricDetailsResponseBodyResources(TeaModel):
     def __init__(
         self,
+        compliance_type: str = None,
         region_id: str = None,
         resource_classification: str = None,
         resource_id: str = None,
@@ -2684,6 +2955,7 @@ class ListEvaluationMetricDetailsResponseBodyResources(TeaModel):
         resource_properties: List[ListEvaluationMetricDetailsResponseBodyResourcesResourceProperties] = None,
         resource_type: str = None,
     ):
+        self.compliance_type = compliance_type
         self.region_id = region_id
         self.resource_classification = resource_classification
         self.resource_id = resource_id
@@ -2704,6 +2976,8 @@ class ListEvaluationMetricDetailsResponseBodyResources(TeaModel):
             return _map
 
         result = dict()
+        if self.compliance_type is not None:
+            result['ComplianceType'] = self.compliance_type
         if self.region_id is not None:
             result['RegionId'] = self.region_id
         if self.resource_classification is not None:
@@ -2724,6 +2998,8 @@ class ListEvaluationMetricDetailsResponseBodyResources(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('ComplianceType') is not None:
+            self.compliance_type = m.get('ComplianceType')
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
         if m.get('ResourceClassification') is not None:
@@ -2838,7 +3114,9 @@ class ListEvaluationResultsRequest(TeaModel):
         account_id: int = None,
         region_id: str = None,
     ):
+        # The Alibaba Cloud account ID of the member. This parameter takes effect only when a multi-account governance maturity check is performed.
         self.account_id = account_id
+        # The region ID.
         self.region_id = region_id
 
     def validate(self):
@@ -2871,7 +3149,9 @@ class ListEvaluationResultsResponseBodyResultsMetricResultsErrorInfo(TeaModel):
         code: str = None,
         message: str = None,
     ):
+        # The error code.
         self.code = code
+        # The error message.
         self.message = message
 
     def validate(self):
@@ -2903,6 +3183,7 @@ class ListEvaluationResultsResponseBodyResultsMetricResultsResourcesSummary(TeaM
         self,
         non_compliant: int = None,
     ):
+        # The number of non-compliant resources.
         self.non_compliant = non_compliant
 
     def validate(self):
@@ -2936,12 +3217,29 @@ class ListEvaluationResultsResponseBodyResultsMetricResults(TeaModel):
         risk: str = None,
         status: str = None,
     ):
+        # The error information.
+        # 
+        # >  This parameter is returned only if the value of `Status` is `Failed`.
         self.error_info = error_info
+        # The end time of the check item. The time is displayed in UTC.
         self.evaluation_time = evaluation_time
+        # The ID of the check item.
         self.id = id
+        # The checked resources.
         self.resources_summary = resources_summary
+        # The rate of the non-compliant resources.
         self.result = result
+        # The risk level. Valid values:
+        # 
+        # *   Error: high risk
+        # *   Warning: medium risk
+        # *   None: no risk
         self.risk = risk
+        # The status of the check item. Valid values:
+        # 
+        # *   Running: The check is in progress.
+        # *   Finished: The check is complete.
+        # *   failed: The check fails.
         self.status = status
 
     def validate(self):
@@ -3001,9 +3299,17 @@ class ListEvaluationResultsResponseBodyResults(TeaModel):
         status: str = None,
         total_score: float = None,
     ):
+        # The end time of the overall check. The time is displayed in UTC.
         self.evaluation_time = evaluation_time
+        # The check results.
         self.metric_results = metric_results
+        # The status of the overall check. Valid values:
+        # 
+        # *   Running: The check is in progress.
+        # *   Finished: The check is complete.
+        # *   failed: The check fails.
         self.status = status
+        # The overall score.
         self.total_score = total_score
 
     def validate(self):
@@ -3053,8 +3359,11 @@ class ListEvaluationResultsResponseBody(TeaModel):
         request_id: str = None,
         results: ListEvaluationResultsResponseBodyResults = None,
     ):
+        # The Alibaba Cloud account ID of the member.
         self.account_id = account_id
+        # The request ID.
         self.request_id = request_id
+        # The check results, including the status of the overall check and the results of check items.
         self.results = results
 
     def validate(self):
@@ -3131,10 +3440,12 @@ class ListEvaluationResultsResponse(TeaModel):
 class ListEvaluationScoreHistoryRequest(TeaModel):
     def __init__(
         self,
+        account_id: int = None,
         end_date: str = None,
         region_id: str = None,
         start_date: str = None,
     ):
+        self.account_id = account_id
         self.end_date = end_date
         self.region_id = region_id
         self.start_date = start_date
@@ -3148,6 +3459,8 @@ class ListEvaluationScoreHistoryRequest(TeaModel):
             return _map
 
         result = dict()
+        if self.account_id is not None:
+            result['AccountId'] = self.account_id
         if self.end_date is not None:
             result['EndDate'] = self.end_date
         if self.region_id is not None:
@@ -3158,6 +3471,8 @@ class ListEvaluationScoreHistoryRequest(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccountId') is not None:
+            self.account_id = m.get('AccountId')
         if m.get('EndDate') is not None:
             self.end_date = m.get('EndDate')
         if m.get('RegionId') is not None: